requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==1.3.0
attrs==24.2.0
beautifulsoup4==4.12.3
behave==1.2.6
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
core-universal4==4.21.0
cryptography==43.0.0
cssselect==1.2.0
decorator==5.1.1
exceptiongroup==1.2.2
execnet==2.1.1
eyes-common==5.29.0
eyes-selenium==5.27.0
fasteners==0.19
filelock==3.16.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
marshmallow==2.21.0
mdurl==0.1.2
mysql-connector-python==9.0.0
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.3
pdbp==1.5.4
platformdirs==4.3.6
pluggy==1.5.0
py==1.11.0
pycparser==2.22
Pygments==2.18.0
pynose==1.5.3
pyOpenSSL==24.2.1
pyotp==2.9.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.3
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-seleniumbase==4.31.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.31.0
rich==13.8.1
sbvirtualdisplay==1.3.0
selenium==3.141.0
self==2020.12.3
setuptools==75.1.0
six==1.16.0
smart-assertions==1.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.3.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==1.26.20
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wsproto==1.2.0
